Tests: improve Plugin module coverage
The most relevant change is the introduction of property-based tests
where relevant. Related to that, a number of support files are
added (e.g. generators for MQTT topics and Astarte payloads) or
reworked (e.g. most of database setup has been moved to a CaseTemplate).
This is the first step in reworking the whole test suite,
therefore existing tests have not been changed.

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>